% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacthexgrid.R
\name{hexagon}
\alias{hexagon}
\title{Create Hexagon Component}
\usage{
hexagon(..., q, r, s)
}
\arguments{
\item{...}{child components}

\item{q, r, s}{`[integer()]` giving axial coordinates, see [readblobgames](https://www.redblobgames.com/grids/hexagons/).}
}
\description{
Create Hexagon Component
}
\seealso{
Other components: \code{\link{hexgrid}},
  \code{\link{layout}}
}
\concept{components}
