% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacthexgrid.R
\name{hexgrid}
\alias{hexgrid}
\title{Create Layout Component}
\usage{
hexgrid(..., width = 1200, height = 800, viewBox = "-50 -50 100 100")
}
\arguments{
\item{...}{child components}

\item{width, height}{`[integer(1)]` giving width and height in pixels.}

\item{viewBox}{`[character(1)]` giving position and dimension of [SVG viewport](https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/viewBox) as four whitespace-separated numbers `"min-x min-y width height"`.}
}
\description{
Create Layout Component
}
\seealso{
Other components: \code{\link{hexagon}},
  \code{\link{layout}}
}
\concept{components}
