% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacthexgrid.R
\name{layout}
\alias{layout}
\title{Create Layout Component}
\usage{
layout(..., size = list(x = 10, y = 10), flat = TRUE, spacing = 1.1,
  origin = list(x = 0, y = 0))
}
\arguments{
\item{...}{child components}

\item{size}{`[list(x = integer(), y = integer())]` giving extent of the grid.}

\item{flat}{`[logical(1)]` giving whether the hexagons should flat (`TRUE`) or pointy-topped (`FALSE`).}

\item{spacing}{`[numeric(1)]` giving the spacing between the hexagons.}

\item{origin}{`[list(x = integer(), y = integer())]` giving the origin of the coordinate system.}
}
\description{
Create Layout Component
}
\seealso{
Other components: \code{\link{hexagon}},
  \code{\link{hexgrid}}
}
\concept{components}
