% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{path}
\alias{path}
\title{Create [Path](https://developer.mozilla.org/en-US/docs/Web/SVG/Element/path) Component}
\usage{
path(start, end)
}
\arguments{
\item{start, end}{`[numeric(2)]` giving the start and end x,y coordinates in the SVG `viewBox` of [HexGrid()].}
}
\description{
Create [Path](https://developer.mozilla.org/en-US/docs/Web/SVG/Element/path) Component
}
