% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{pattern}
\alias{pattern}
\title{Create [Pattern](https://developer.mozilla.org/en-US/docs/Web/SVG/Element/pattern) Component}
\usage{
pattern(..., id, link)
}
\arguments{
\item{...}{child components}

\item{id}{`[character(1)]`}

\item{link}{`[character(1)]`}
}
\description{
Create [Pattern](https://developer.mozilla.org/en-US/docs/Web/SVG/Element/pattern) Component
}
\seealso{
Other components: \code{\link{hexagon}},
  \code{\link{hexgrid}}, \code{\link{layout}},
  \code{\link{text}}
}
\concept{components}
