% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacthexgrid.R
\name{reacthexgrid-shiny}
\alias{reacthexgrid-shiny}
\alias{reacthexgridOutput}
\alias{renderReacthexgrid}
\alias{reacthexgrid_html}
\title{Shiny bindings for reacthexgrid}
\usage{
reacthexgridOutput(outputId, width = "100\%", height = "400px")

renderReacthexgrid(expr, env = parent.frame(), quoted = FALSE)

reacthexgrid_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a reacthexgrid}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using reacthexgrid within Shiny
applications and interactive Rmd documents.
}
