% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacthexgrid.R
\name{reacthexgrid}
\alias{reacthexgrid}
\title{Create hexagon grid}
\usage{
reacthexgrid(width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\description{
Wraps the [react-hexgrid](https://www.npmjs.com/package/react-hexgrid) library.
}
\concept{wrappers}
