% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{text}
\alias{text}
\title{Create Text Component}
\usage{
text(x = NULL)
}
\arguments{
\item{x}{`[character(1)]` giving a string to be shown inside the hexagons.}
}
\description{
Create Text Component
}
\seealso{
Other components: \code{\link{hexagon}},
  \code{\link{hexgrid}}, \code{\link{layout}}
}
\concept{components}
