% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{hexagon}
\alias{hexagon}
\title{Create Hexagon Component}
\usage{
hexagon(..., q, r, s)
}
\arguments{
\item{...}{child components}

\item{q, r, s}{`[integer()]` giving axial coordinates, see [readblobgames](https://www.redblobgames.com/grids/hexagons/).
The axial coordinates are unitless integers; they need not correspond the SVG coordinate system of [Hexagon()].}
}
\description{
Create Hexagon Component
}
\seealso{
Other components: \code{\link{hexgrid}},
  \code{\link{layout}}, \code{\link{text}}
}
\concept{components}
