% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{hexgrid}
\alias{hexgrid}
\title{Create Layout Component}
\usage{
hexgrid(..., width = "100\%", height = "100\%",
  viewBox = "-50 -50 100 100")
}
\arguments{
\item{...}{child components}

\item{width, height}{`[integer(1)]` giving width and height of [outermost svg element](https://developer.mozilla.org/en-US/docs/Web/SVG/Element/svg) viewport, expressed in percentage of parent element or CSS length units.}

\item{viewBox}{`[character(1)]` giving the cartesian dimensions of the [SVG viewport](https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/viewBox) as four whitespace-separated numbers `"min-x min-y width height"`.}
}
\description{
Create Layout Component
}
\seealso{
Other components: \code{\link{hexagon}},
  \code{\link{layout}}, \code{\link{text}}
}
\concept{components}
